==> 00_data_processing/04_prepare_data_for_OSF.ipynb <==

import json
import pandas as pd
import re
from pyprojroot import here


# Define the file path
round1 = 'data/recoded_transcripts/main-at-2024-05-28-recoded.json'
round2 = 'data/individual_policy_data/policy-recoding-2025-05-21-complete.json'
round2_binary = '00_data_processing/output/policy_category_df_binary.csv'

# Load the JSON files
with open(here(round1), 'r') as file:
    data = json.load(file)

with open(here(round2), 'r') as file:
    data_round2 = json.load(file)

round2_binary = pd.read_csv(here(round2_binary))



# # Convert the list of dictionaries into a DataFrame
df = pd.DataFrame(data)

df_round2 = pd.DataFrame(data_round2)



round2_binary.columns

df_round2.columns

df

df_round2


# Expand the 'meta_info' column into separate columns
meta_info_df = pd.json_normalize(df["meta_info"])

# Concatenate the original DataFrame with the expanded meta_info columns
df = pd.concat([df.drop(columns=["meta_info"]), meta_info_df], axis=1)



# Extract the LexisNexis ID using regex
df["text"] = df["text"].str.replace(r"\s+", " ", regex=True).str.strip()
df["LexisNexis_id"] = df["text"].str.extract(r"Transcript:\s*([A-Za-z0-9.]+)", flags=re.DOTALL)

df_round2["text"] = df_round2["text"].str.replace(r"\s+", " ", regex=True).str.strip()
df_unique = df[['internal_id', 'source', 'date', 'LexisNexis_id']].drop_duplicates(subset='internal_id')

# Now merge
df_round2 = pd.merge(
    df_round2,
    df_unique,
    on='internal_id',
    how='left'
)



df_to_export_round1 = df[['internal_id', 'LexisNexis_id', 'source', 'date', 'annotator', 'climate_change', 'attitude', 'policy']]

merged_df_round2 = pd.merge(df_round2, round2_binary, on=['internal_id', 'annotator'], suffixes=('', '_binary'))

keep_columns_round2 = [
    'internal_id', 'LexisNexis_id', 'source', 'date', 'annotator',
    'Paris Agreement', 'Green New Deal', 'Executive action', 'Emergency declaration',
    '(De)regulation and laws', 'Renewable energy', 'Emission reduction',
    'Oil and gas industry', 'Other climate policy'
]

df_to_export_round2 = merged_df_round2[keep_columns_round2]

df_to_export_round1.to_csv('output/OSF/annotation_data_round1.csv', index=False)
df_to_export_round2.to_csv('output/OSF/annotation_data_round2.csv', index=False)
